This commit adds the sticky footer frame.  It DOES NOT add the footer.
If we need a footer (do we?  Probably not) we need to create a render/Footer.php to create it.

This also modifies a few files that I came across to fit with a flexible design.
Also, I've added the local.css file for site-wide (but not global) styling.  sticky footer stuff is in there.